продолжаю бороться с исключениями
есть продвижения... у каждого репозитория есть метод проверки записи с определёнными полями или одним полем. по номеру записи он имеется по умолчанию, а осталиные нужно создавать в соответствующем репозитории, объявив ему нужное имя...
всё за выходной думаю завершить...
И просьба БОЛЬШАЯ... не трогать уровень логирования... он мне СИЛЬНО помогает в написании программы...
я не плодил исключения для каждого красса и я их вызываю на уровне "модель", так как на уровне "репозиторий" это или тяжело или не возможно, пускать его на ружу я не хочу и считаю это грубой ошибкой... такой как передавать между BackEnd и FrontEnd сущности, хранящиеся в нашей БД.

+++ MishpahugHaifa/src/main/java/application/models/event/EventModel.java
@@ -7,6 +7,7 @@ import java.util.stream.Collectors;
 
 import javax.transaction.Transactional;
 
+import application.exceptions.NotFoundEntityException;
 import application.utils.converter.IUpdates;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -53,6 +54,7 @@ public class EventModel implements IEventModel {
 
 	@Override
 	public Set<EventEntity> getAllByUser(Integer userId) {// TODO: rename to getSubscriptionsByUser
+		if (!userRepository.existsById(userId)) throw new NotFoundEntityException("");
 		UserEntity userEntity = userRepository.getOne(userId);
 		Set<SubscriptionEntity> subscriptions = userEntity.getSubscriptions();
 		return subscriptions.stream().map(s -> s.getEvent()).collect(Collectors.toSet());
@@ -60,6 +62,7 @@ public class EventModel implements IEventModel {
 
 	@Override
 	public Set<UserEntity> getAllSubscribed(Integer eventId) {// TODO: rename to get GuestsByEvent
+		if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
 		EventEntity eventEntity = eventRepository.getOne(eventId);
 		Set<SubscriptionEntity> subscriptions = eventEntity.getSubscriptions();
 		return subscriptions.stream().map(s -> s.getGuest()).collect(Collectors.toSet());
@@ -77,6 +80,7 @@ public class EventModel implements IEventModel {
 
 	@Override
 	public EventEntity update(Integer eventId, HashMap<String, String> data){
+		if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
 		EventEntity eventEntity = eventRepository.getOne(eventId);
 		updates.updateEvent(eventEntity, data);
 		return eventRepository.save(eventEntity);
@@ -84,6 +88,7 @@ public class EventModel implements IEventModel {
 
 	@Override
 	public EventEntity delete(Integer eventId){ // throws if not in deletion queue
+		if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
 		EventEntity eventEntity = eventRepository.getOne(eventId);
 		eventRepository.delete(eventEntity);
 		return eventEntity;
@@ -95,8 +100,9 @@ public class EventModel implements IEventModel {
 	}
 
 	@Override
-	public EventEntity getById(Integer id){
-		return eventRepository.getOne(id);
+	public EventEntity getById(Integer eventId){
+		if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
+		return eventRepository.getOne(eventId);
 	}
 
 	@Override
@@ -111,18 +117,24 @@ public class EventModel implements IEventModel {
 
 	@Override
 	public EventEntity subscribe(Integer eventId, Integer userId){
+		if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
+		if (!userRepository.existsById(userId)) throw new NotFoundEntityException("");
 		SubscriptionHandler handler = new SubscriptionHandler(eventId, userId);
 		return handler.subscribe();
 	}
 
 	@Override
 	public EventEntity unsubscribe(Integer eventId, Integer userId){
+		if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
+		if (!userRepository.existsById(userId)) throw new NotFoundEntityException("");
 		SubscriptionHandler handler = new SubscriptionHandler(eventId, userId);
 		return handler.unsubscribe();
 	}
 
 	@Override
 	public EventEntity deactivateSubscription(Integer eventId, Integer userId){
+		if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
+		if (!userRepository.existsById(userId)) throw new NotFoundEntityException("");
 		SubscriptionHandler handler = new SubscriptionHandler(eventId, userId);
 		return handler.deactivate();
 	}
@@ -146,6 +158,8 @@ public class EventModel implements IEventModel {
 		}
 
 		private void load(){
+			if (!eventRepository.existsById(eventId)) throw new NotFoundEntityException("");
+			if (!userRepository.existsById(userId)) throw new NotFoundEntityException("");
 			eventEntity = eventRepository.getOne(eventId);
 			userEntity = userRepository.getOne(userId);
 			EventGuestId subscriptionKey = new EventGuestId(userEntity.getId(), eventEntity.getId());
